Re-implement object rendering using a depth-sorted approach.
This moves the object rendering logic into ObjectRenderer. This makes
GameRenderer a bit smaller.

There are some rendering logic decisions that haven't been brought
back yet since they may be better placed elsewhere.

+++ rwengine/include/data/ObjectData.hpp
@@ -45,7 +45,7 @@ struct ObjectData : public ObjectInformation
 
 	std::string modelName;
 	std::string textureName;
-	size_t numClumps;
+	uint8_t numClumps;
 	float drawDistance[3];
 	int32_t flags;
 	bool LOD;
